pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rscovlm"
version = "0.0.1"
description = "A training codebase of multimodal language model for geoscience remote sensing."
authors = [
    { name = "Li-Qingyun", email = "962537281@qq.com" },
    { name = "Shuran Ma", email = "dadaochengqi@163.com" },
    { name = "Xue Yang", email = "yangxue0827@126.com" },
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch",
    "torchvision",
    "transformers",
    "peft==0.14.0",
    "datasets",
    "huggingface_hub",
    "timm",
    "deepspeed<=0.16.9",
    "accelerate>=0.29.1",
    "numpy<2",
    "einops",
    "tensorboard",
    "scipy", 
    "scikit-learn", 
    "rapidfuzz", 
    "openmim", 
    "fire",
    "scikit-image",
    "Levenshtein",
    "xmltodict",
    "pycocoevalcap",
    "shapely",
    "binpacking", 
    "python-magic", 
    "qwen_vl_utils==0.0.11", 
    "simplejson", 
    "nltk"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[tool.setuptools.packages.find]
include = ["rscovlm*"]
exclude = [
    "checkpoints*",
    "playground*",
    "scripts*",
    "scripts_sh*",
    "scripts_py*",
    "tests*",
    "wandb*",
]

[tool.wheel]
exclude = [
    "checkpoints*",
    "playground*",
    "scripts*",
    "scripts_sh*",
    "scripts_py*",
    "tests*",
    "wandb*",
]
